% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitmerits.R
\name{colGI}
\alias{colGI}
\title{Gini index and optimal threshold for columns of scores}
\usage{
colGI(scores, y)
}
\arguments{
\item{scores}{A matrix of scores, each column is one variant}

\item{y}{A logical decision}
}
\value{
A matrix of results; cols correspond to variants, rows to respectively GI and optimal threshold
}
\description{
Information gain and optimal threshold for columns of scores
}
\examples{
colGI(cbind(iris[,-5],rep(77,150)),iris$Species=='versicolor')
}
