% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{createTubeset}
\alias{createTubeset}
\title{Convert a set of FCS files into a tubeset}
\usage{
createTubeset(
  fNamesInTubes,
  dirOut,
  reduce = 1,
  removeColumns = NULL,
  verbose = FALSE
)
}
\arguments{
\item{fNamesInTubes}{A list of character vectors; n-th element corresponds to n-th tube, its m-th element is a path to an FCS file describing m-th object}

\item{dirOut}{Path to write output files; may not exist, function will try to recursively create it}

\item{reduce}{Only extract random sample of \code{reduce * 100\%} of events. Set to 1 to disable any manipulation}

\item{removeColumns}{A numeric vector of parameter indices which should be dropped from the converted set; useful to remove nonsense data like event ID or time}

\item{verbose}{Set to \code{TRUE} to enable process progress reporting}
}
\value{
Invisible \code{NULL}; the important side effect is a set of files \code{<dirOut>/tube<N>.bin} holding the converted data, which can be loaded with \code{loadTubeset} function.
}
\description{
This function allows one to create a >>tubeset<< (a specific form of cytometry data on which flowForest can operate) from a set of FCS files
}
\note{
You don't have to create separate tubesets for training and test data; this can be specified as a mask given to the flowForest function.

Converter will overwrite existing \code{tube<N>.bin} files, yet it is adviced to use empty paths (for instance, N-tube set written over N+1-tube one will leave one old tube file, leading to either load failure or unexpected results).
}
