% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert.R
\name{createTubsetRaw}
\alias{createTubsetRaw}
\title{Convert a set of data.frames into a tubeset}
\usage{
createTubsetRaw(dfList, dirOut, verbose = FALSE)
}
\arguments{
\item{dfList}{A list of data.frames or a list of lists of data.frames. In the first case, it is assumed that there is a single tube, otherwise that each top-levl list corresponds to a tube. Within a tube, each data.frame corresponds to an object, and its rows to events recorded for this object. Colums of a data.frame represent the recorded event properties, thus their count must be equal for all data.frames from a tube.}

\item{dirOut}{Path to write output files; may not exist, function will try to recursively create it}

\item{verbose}{Set to \code{TRUE} to enable process progress reporting}
}
\value{
Invisible \code{NULL}; the important side effect is a set of files \code{<dirOut>/tube<N>.bin} holding the converted data, which can be loaded with \code{\link{loadTubeset}} function.
}
\description{
This function allows one to create a >>tubeset<< (a specific form of cytometry data on which flowForest can operate) from a set of \code{data.frame}s.
}
\note{
You don't have to create separate tubesets for training and test data; this can be specified as a mask given to the flowForest function.

Converter will overwrite existing \code{tube<N>.bin} files, yet it is adviced to use empty paths (for instance, N-tube set written over N+1-tube one will leave one old tube file, leading to either load failure or unexpected results).
}
