% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{dotGate}
\alias{dotGate}
\title{Apply a gate on a tubeset}
\usage{
dotGate(set, gate, mask = 1:set[2])
}
\arguments{
\item{set}{Tubeset}

\item{gate}{Gate}

\item{mask}{A numeric mask specifying on which objects gate shall be applied.}
}
\value{
Matrix of a dim number of objects x number of parameters+1.
Rows correspond to objects, first column to a number of events inside gate,
n+1 column to average value of a parameter n over events inside gate.
}
\description{
Apply a gate on a tubeset
}
