% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{extractEvents}
\alias{extractEvents}
\title{Extract events}
\usage{
extractEvents(set, tube, events)
}
\arguments{
\item{set}{Object representing data}

\item{tube}{Tube index (must be a single number)}

\item{events}{Indices of the selected events from a flowset composed of all flow frames in a given tube}
}
\value{
Data frame, rows are events in a requested order, columns are parameters as in the original flow frame.
}
\description{
Read a given subset of events from a single tube. Useful for visualisation.
}
