% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{extractGatesFromflowForest}
\alias{extractGatesFromflowForest}
\title{Extract all gates used to build flowForest}
\usage{
extractGatesFromflowForest(flowForest, nOfTube)
}
\arguments{
\item{flowForest}{Object of a class \code{flowForest}, as that created by the function flowForest.}

\item{nOfTube}{Number of tube in tubeset on which flowForest was build; only gates on this tube will be extracted.}
}
\description{
Extract all gates used to build flowForest
}
