% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{flowForest}
\alias{flowForest}
\title{Train a flowForest model}
\usage{
flowForest(
  x,
  y,
  trainMask = 1:x[2],
  testMask = (1:x[2])[!(1:x[2]) \%in\% trainMask],
  nOfTrees = 100,
  gateTries = 5,
  gateMode = "pseudosphere",
  asymmetric = TRUE,
  keepForest = TRUE,
  verbose = TRUE,
  relaxedGateAccept = FALSE
)
}
\arguments{
\item{x}{Flow cytometry measurements, in a form of an object of the class \code{tubeset}, as that created by the function \code{loadTubeset}.}

\item{y}{Decision; must be a logical vector of a same length as the number of objects in \code{x}. Unknown decisions may be marked with \code{NA}; they can't be in the training set, though}

\item{trainMask}{A vector of indices of objects in tubeset which are to be assumed to reside in the trainig set. By default the whole set}

\item{testMask}{A vector of indices of objects in tubeset which are to be assumed to reside in the test set. By default those not in \code{trainMask}}

\item{nOfTrees}{Number of trees to grow}

\item{gateTries}{Number of gates to try at each split; this does not count gates rejected due to insufficient event fraction}

\item{gateMode}{Mode of the gate generation; only \code{pseudosphere} is currently supported. See \code{\link{makeRandomGate}} for more information.}

\item{asymmetric}{If \code{TRUE}, only gates which hold on average more events for the \code{TRUE} class will be taken into account. This is useful to force flowForest to focus on event populations characteristic for the \code{TRUE} class.}

\item{keepForest}{Shall the forest structure be saved? This is required for importance and predicting new data, but requires more memory.}

\item{verbose}{If \code{TRUE}, performance scores of \code{flowForest} counted on test set and OOB will be printed after growing each tree. Otherwise, only after building whole \code{flowForest}.}

\item{relaxedGateAccept}{Accept gates with some event, not only at least one event for all objects.}
}
\value{
Object of a class \code{flowForest}
}
\description{
Train a flowForest model
}
