% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{gateSize}
\alias{gateSize}
\title{Number of events in gate on given tubeset}
\usage{
gateSize(gate, tubeset, cellsToCheck)
}
\arguments{
\item{gate}{Object of a class \code{gate}, as that extracted from flowForest by function \code{extractGatesFromflowForest}.}

\item{tubeset}{Flow cytometry measurements, in a form of a tubeset, as that created by the function \code{loadTubeset}.}

\item{cellsToCheck}{Vector of integers indicating, to which cells in tube, on which the gate is made, narrow counting. By default all cells in this tube.}
}
\description{
Number of events in gate on given tubeset
}
