% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.R
\name{importance}
\alias{importance}
\title{Identify predictive cell populations}
\usage{
importance(
  tubeset,
  flowForest,
  nOfTube = 1,
  threshold = 0.95,
  transparency = 0.1,
  sizePenalty = TRUE,
  additionalPlots = FALSE
)
}
\arguments{
\item{tubeset}{Flow cytometry measurements, in a form of a tubeset, as that created by the function \code{loadTubeset}.}

\item{flowForest}{Object of a class \code{flowForest}, trained by the function flowForest on \code{tubeset}.}

\item{nOfTube}{Number of tube in tubeset on which flowForest was build; only gates on this tube will be extracted.}

\item{threshold}{A vector of indices of objects in tubeset which are to be assumed to reside in the test set. By default those not in \code{trainMask}.}

\item{transparency}{Transparency of events on the scatterplot produced by the function.}

\item{sizePenalty}{Shall gates comprising many events be considered less important?}

\item{additionalPlots}{If \code{FALSE}, only histogram of scores of events and scatterplot with important events will be printed. Otherwise, additionally barplot with frequency of usage of each criterion on gates in splits will be printed too.}
}
\value{
Data frame with coordinates of events from the tubeset (all columns except last two), it's scores of importance (the penultimate column) and logical vector indicating, if they were considered as important or not (the last column).
}
\description{
Identify predictive cell populations
}
