% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{inGate}
\alias{inGate}
\title{Check whether certain events are enclosed by a gate}
\usage{
inGate(set, gate, idx)
}
\arguments{
\item{set}{Tubeset}

\item{gate}{Gate}

\item{idx}{Indices of the selected events, similar to what \code{\link{extractEvents}} expects}
}
\value{
Boolean vector of a same size and orfer as \code{idx}. Holds \code{TRUE}s for events inside the given gate.
}
\description{
Check whether certain events are enclosed by a gate
}
