% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{loadTubeset}
\alias{loadTubeset}
\title{Load tubset}
\usage{
loadTubeset(path)
}
\arguments{
\item{path}{A path to a directory with tube<N>.bin files representing tubeset (i.e. output of \code{createTubeset})}
}
\value{
Object representing data
}
\description{
Load a conveted set of FCS files
}
\note{
Tubeset will be freed automatically by R; loading is done by
memory mapping, hence loading the same data from other process on the same
machine will not consume additional memory.
Also, flowForest may load datasets larger than available RAM; still this
may impose a substantial slowdown.
}
