% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{makeRandomGate}
\alias{makeRandomGate}
\title{Generate a random gate}
\usage{
makeRandomGate(set, gateMode)
}
\arguments{
\item{set}{Flow cytometry measurements, in a form of an object of the class \code{tubeset}, as that created by the function \code{loadTubeset}.}

\item{gateMode}{Mode of the gate generation; only \code{pseudosphere} currently supported.}
}
\value{
Object of a class \code{gate}, random gate of a mode d.
}
\description{
This function generates a random gate on a randomly selected tube; flowForest builds many such gates to find ones which can be useful for separating classes. In principle, there can be many ways in which such gate is generated, but the current version of flowForest only supports one, \code{pseudosphere}.
}
\details{
In the \code{pseudosphere} mode, the code first selects a random tube. Then, a random event from the tube, which will be the geometrical centre of the final gate. Finally, the gate is expanded in all dimensions, by a random fraction of the range between 5-th and 95-th percentile in a certain dimension.
}
