% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forest.R
\name{predict.flowForest}
\alias{predict.flowForest}
\title{Predict method for flowForest objects}
\usage{
\method{predict}{flowForest}(object, x, testMask = 1:x[2], type = "response", ...)
}
\arguments{
\item{object}{\code{flowForest} object (as that created by the function \code{flowForest}), that will be used for prediction.}

\item{x}{Test data, flow cytometry measurements in a form of a \code{tubeset}, as that created by the function \code{loadTubeset}.}

\item{testMask}{A vector of indices of objects in tubeset which are to be assumed to reside in the test set. By default the whole set.}

\item{type}{\code{"score"} or \code{"class"} - types of returned value. In first case score from trees voting for each object is returned; otherwise - the most frequent class in voting.}

\item{...}{Ignored}
}
\value{
If \code{type == "class"} - predicted class of each object in tubeset is returned; if \code{type == "score"} - score is returned - a real number from between 0 and 1, obtained from voting.
}
\description{
This function predicts classes of new objects with given \code{flowForest} object.
}
