% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.flowForest}
\alias{print.flowForest}
\title{Print flowForest}
\usage{
\method{print}{flowForest}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{flowForest}}

\item{...}{Ignored}
}
\value{
invisibly \code{x}
}
\description{
Prints the flowForest object.
}
