% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print.gate}
\alias{print.gate}
\title{Print gate}
\usage{
\method{print}{gate}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{gate}}

\item{...}{Ignored}
}
\value{
invisibly \code{x}
}
\description{
Prints the gate object.
}
