% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/raw.R
\name{print.tubeset}
\alias{print.tubeset}
\title{Print tubset}
\usage{
\method{print}{tubeset}(x, ...)
}
\arguments{
\item{x}{Object of a class \code{tubeset}}

\item{...}{Ignored}
}
\value{
invisibly \code{x}
}
\description{
Prints the tubeset object.
}
